+++ jsonjsdb/test/db/gdf9898fds/email.json.js
@@ -1,1 +1,1 @@
-jsonjs.data['email'] = [['id','name','user_id','admin_id'],['email_1','email 1',1,2],['email_2','email 2',2,3]]
+jsonjs.data['email'] = [['id','name','user_id','admin_user_id','partner_user_id'],['email_1','email 1',1,2,3],['email_2','email 2',2,3,4]]