+++ jsonjsdb/test/db/gdf9898fds/user.json.js
@@ -1,1 +1,1 @@
-jsonjs.data['user'] = [['id','name','doc_ids'],[1,'user 1','3, 4, 5'],[2,'user 2'],[3,'user 3'],[4,'user 4'],[5,'user 5']]
+jsonjs.data['user'] = [['id','name','doc_ids','source_user_ids'],[1,'user 1','3, 4, 5'],[2,'user 2',null,'1'],[3,'user 3',null,'1, 2'],[4,'user 4'],[5,'user 5']]